[update]page loading

+++ cron_flask_api.py
@@ -65,10 +65,12 @@ def add_cron():
     # 判断是否已经存在，已存在则不用新建
     for job in my_cron:
         if job.command.strip() == auto_cmd.strip():
-            job.enable(False)
-            my_cron.write()
-            create_res = True
-            break
+            job_frequency = str(job).split('root')[0].strip()
+            if job_frequency == frequency:
+                job.enable(False)
+                my_cron.write()
+                create_res = True
+                break
     if not create_res:
         job = my_cron.new(command=auto_cmd, user='root')
         job.setall(frequency.strip())
@@ -85,6 +87,7 @@ def del_cron():
     errcode = 0
     msg = 'ok'
     auto_cmd = request.form.get('auto_cmd')
+    frequency = request.form.get('frequency').strip()
 
     # 在机器上暂停任务
     my_cron = CronTab(tabfile='/etc/crontab', user=False)
@@ -93,10 +96,11 @@ def del_cron():
     print auto_cmd
     for job in my_cron:
         if job.command == auto_cmd:
-            job.enable(False)
-            print 'modify_cron----disable---done'
-            my_cron.write()
-            break
+            job_frequency = str(job).split('root')[0].strip()
+            if job_frequency == frequency:
+                job.enable(False)
+                my_cron.write()
+                break
 
     data = dict(code=errcode, msg=msg)
     return jsonify(data)
@@ -118,11 +122,13 @@ def multi_del_cron():
         auto_cmd = cron_obj.auto_cmd.strip()
         for job in my_cron:
             if job.command == auto_cmd:
-                job.enable(False)
-                my_cron.write()
-                # DB中删除
-                cron_obj.delete()
-                break
+                job_frequency = str(job).split('root')[0].strip()
+                if job_frequency == cron_obj.frequency:
+                    job.enable(False)
+                    my_cron.write()
+                    # DB中删除
+                    cron_obj.delete()
+                    break
 
     data = dict(code=errcode, msg=msg)
     return jsonify(data)
@@ -148,10 +154,12 @@ def start_cron():
         print auto_cmd
         for job in my_cron:
             if job.command == auto_cmd:
-                job.enable()
-                print 'start_cron----enable---done'
-                my_cron.write()
-                break
+                job_frequency = str(job).split('root')[0].strip()
+                if job_frequency == crontab_obj.frequency:
+                    job.enable()
+                    print 'start_cron----enable---done'
+                    my_cron.write()
+                    break
 
         # 修改数据库中cmd状态
         user_obj = User.objects.get(username=username)
@@ -179,14 +187,16 @@ def pause_cron():
         # 修改机器上crontab状态为启动
         my_cron = CronTab(tabfile='/etc/crontab', user=False)
         auto_cmd = crontab_obj.auto_cmd.strip()
-        print 'start_cron---auto_cmd : '
+        print 'pause_cron---auto_cmd : '
         print auto_cmd
         for job in my_cron:
             if job.command == auto_cmd:
-                job.enable(False)
-                print 'start_cron----enable---done'
-                my_cron.write()
-                break
+                job_frequency = str(job).split('root')[0].strip()
+                if job_frequency == crontab_obj.frequency:
+                    job.enable(False)
+                    print 'pause_cron----enable---done'
+                    my_cron.write()
+                    break
 
         # 修改数据库中cmd状态
         user_obj = User.objects.get(username=username)
